+++ hpOneView/oneview_client.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2019] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -107,6 +107,9 @@ from hpOneView.resources.settings.appliance_device_snmp_v3_users import Applianc
 from hpOneView.resources.settings.appliance_node_information import ApplianceNodeInformation
 from hpOneView.resources.settings.appliance_time_and_locale_configuration import ApplianceTimeAndLocaleConfiguration
 from hpOneView.resources.settings.versions import Versions
+from hpOneView.resources.hypervisors.hypervisor_managers import HypervisorManagers
+from hpOneView.resources.security.certificates_server import CertificatesServer
+from hpOneView.resources.hypervisors.hypervisor_cluster_profiles import HypervisorClusterProfiles
 
 ONEVIEW_CLIENT_INVALID_PROXY = 'Invalid Proxy format'
 
@@ -198,6 +201,8 @@ class OneViewClient(object):
         self.__backups = None
         self.__login_details = None
         self.__licenses = None
+        self.__hypervisor_managers = None
+        self.__certificates_server = None
 
     @classmethod
     def from_json_file(cls, file_name):
@@ -1171,3 +1176,33 @@ class OneViewClient(object):
         if not self.__licenses:
             self.__licenses = Licenses(self.__connection)
         return self.__licenses
+
+    @property
+    def hypervisor_managers(self):
+        """
+        Gets the Hypervisor Managers API client.
+
+        Returns:
+            HypervisorManagers
+        """
+        return HypervisorManagers(self.__connection)
+
+    @property
+    def certificates_server(self):
+        """
+        Gets the Certificates Server API client.
+
+        Returns:
+            Server Certificate:
+        """
+        return CertificatesServer(self.__connection)
+
+    @property
+    def hypervisor_cluster_profiles(self):
+        """
+        Gets the Hypervisor Cluster Profiles API client.
+
+        Returns:
+            Hypervisor Cluster Profiles:
+        """
+        return HypervisorClusterProfiles(self.__connection)